+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 27
     buildToolsVersion '26.0.2'
     defaultConfig {
-        applicationId "com.example.tomru.breakout"
+        applicationId "com.dreamteam.blocks"
         minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
