+++ app/src/main/java/com/dreamteam/blocks/Brick.java
@@ -1,4 +1,4 @@
-package com.example.tomru.breakout;
+package com.dreamteam.blocks;
 
 import android.graphics.RectF;
