+++ app/src/main/res/layout/activity_breakout_game.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.tomru.breakout.BreakoutGame">
+    tools:context="com.dreamteam.blocks.BreakoutGame">
 
     <TextView
         android:layout_width="wrap_content"
