+++ app/src/main/res/layout/activity_start.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.tomru.breakout.StartActivity">
+    tools:context="com.dreamteam.blocks.StartActivity">
 
     <!-- Login progress -->
     <ProgressBar
