--- app/src/main/res/xml/pref_headers.xml
@@ -1,6 +1,0 @@
-<preference-headers>
-
-    <!-- These settings headers are only used on tablets. -->
-
-
-</preference-headers>
